packet handling, acks, bones of application layer

+++ embedded/atsams70-tinyrouter/atsams70-tinyrouter/src/node.h
@@ -5,6 +5,6 @@
 #define MYADDRESS 2;
 
 uint8_t LUT[1024][4];
-uint16_t myAddress;
+uint8_t myAddress;
 
 #endif
